crescendo_health/requirements.txt
asgiref==3.5.0
Authlib==0.15.5
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
Flask==2.0.3
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
pycparser==2.21
python-dotenv==0.19.2
requests==2.27.1
urllib3==1.26.8
Werkzeug==2.0.3
